Improved artwork detection
Added support for an "art" folder (within the main application folder)
into which the user can place a cover art set in .png or .jpg format.
When the user adds a game, hakchi2 will look in the "art" folder for a
cover art filename that matches the ROM filename. If hakchi2 finds a
match, it will automatically use that cover art for the added game.