[Fix] Automatically load cover art that is stored within the same folder as an identically-named game archive (.7z/.zip/.rar)
Looks in the folder, housing the game archive (.7z/.zip/.rar), for an
identically-named cover art file (.png/.jpg). If found, the cover art is
automatically loaded. Previously, the game could not be within an
archive for the cover art to be found within that folder.

This has been tested with archives (.7z/.zip/.rar) containing the
following file types:
- .32X (Sega 32X)
- .fds (Famicom Disk System)
- .gb (Nintendo Game Boy)
- .gbc (Nintendo Game Boy Color)
- .gg (Sega Game Gear)
- .md (Sega Genesis / Mega Drive)
- .nes (Nintendo Entertainment System / Famicom)
- .pce (NEC PC Engine / TurboGrafx 16)
- .rom (Microsoft MSX / MSX2)
-  sfc & .smc (Super Nintendo / Super Famicom),
- .sms (Sega Master System)